requirements.txt
appnope==0.1.0
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
certifi==2018.10.15
chardet==3.0.4
decorator==4.4.0
defusedxml==0.6.0
Django==2.0
django-crispy-forms==1.7.2
django-filter==2.1.0
django-multiselectfield==0.1.8
djangorestframework==3.9.4
entrypoints==0.3
facebook-sdk==3.1.0
html5lib==1.0.1
httpie==1.0.2
idna==2.8
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
mysqlclient==1.4.2.post1
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.0
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pipenv==2018.11.26
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.4.2
PyMySQL==0.9.3
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2018.7
pyzmq==18.1.0
qtconsole==4.5.3
requests==2.21.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.24.1
virtualenv==16.6.0
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
